pyproject.toml
[tool.poetry]
name = "repl_python3_Masaya-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.0"
pychalk = "^2.0.1"
youtube_dl = "^2021.6.6"
PyNaCl = "^1.4.0"
pafy = "^0.5.5"
ffmpeg = "^1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
